Fix yeelight recorder warning
